+++ packages/cli/src/lib/CodeGenerator.ts
@@ -16,21 +16,16 @@ import {
   AnyManifest,
   SchemaComposer,
   intlMsg,
+  resetDir,
 } from "./";
 
-import { TypeInfo } from "@web3api/schema-parse";
-import {
-  bindSchema,
-  BindLanguage,
-  GenerateBindingFn,
-} from "@web3api/schema-bind";
+import { BindLanguage, GenerateBindingFn } from "@web3api/schema-bind";
 import { writeDirectorySync } from "@web3api/os-js";
 import path from "path";
-import fs, { readFileSync } from "fs";
+import { readFileSync } from "fs";
 import * as gluegun from "gluegun";
 import { Ora } from "ora";
 import Mustache from "mustache";
-import rimraf from "rimraf";
 
 export interface CodeGeneratorConfig {
   outputDir: string;
@@ -46,10 +41,10 @@ export class CodeGenerator {
 
   constructor(private _config: CodeGeneratorConfig) {}
 
-  public async generate(config?: Record<string, unknown>): Promise<boolean> {
+  public async generate(): Promise<boolean> {
     try {
       // Compile the API
-      await this._generateCode(config);
+      await this._generateCode();
 
       return true;
     } catch (e) {
@@ -58,7 +53,7 @@ export class CodeGenerator {
     }
   }
 
-  private async _generateCode(config?: Record<string, unknown>) {
+  private async _generateCode() {
     const { schemaComposer, project } = this._config;
 
     const run = async (spinner?: Ora) => {
@@ -87,7 +82,7 @@ export class CodeGenerator {
       }
 
       // Make sure the output dir is reset
-      this._resetDir(this._config.outputDir);
+      resetDir(this._config.outputDir);
 
       // Get the fully composed schema
       const composed = await schemaComposer.getComposedSchemas();
@@ -125,12 +120,12 @@ export class CodeGenerator {
         }
 
         const output = await generateBinding({
+          projectName: await project.getName(),
           modules: [
             {
               name: "custom",
               typeInfo,
               schema: this._schema || "",
-              config: config || {},
               outputDirAbs: this._config.outputDir,
             },
           ],
@@ -146,21 +141,18 @@ export class CodeGenerator {
           );
         }
       } else {
-        const output = bindSchema({
-          modules: [
-            {
-              name: "combined",
-              typeInfo: composed.combined?.typeInfo as TypeInfo,
-              schema: composed.combined?.schema as string,
-              config,
-              outputDirAbs: this._config.outputDir,
-            },
-          ],
-          bindLanguage,
-        });
+        const output = await project.generateSchemaBindings(
+          composed,
+          this._config.outputDir
+        );
 
+        // Output the bindings
         for (const module of output.modules) {
-          writeDirectorySync(this._config.outputDir, module.output);
+          writeDirectorySync(module.outputDirAbs, module.output);
+        }
+
+        if (output.common) {
+          writeDirectorySync(output.common.outputDirAbs, output.common.output);
         }
       }
     };
@@ -177,14 +169,6 @@ export class CodeGenerator {
         }
       );
     }
-  }
-
-  private _resetDir(dir: string) {
-    if (fs.existsSync(dir)) {
-      rimraf.sync(dir);
-    }
-
-    fs.mkdirSync(dir, { recursive: true });
   }
 
   private _generateTemplate(