+++ packages/cli/src/lib/helpers/client.ts
@@ -17,8 +17,10 @@ export function getSimpleClient(config: SimpleClientConfig): Web3ApiClient {
     plugins.push({
       uri: "w3://ens/ens.web3api.eth",
       plugin: ensPlugin({
-        addresses: {
-          testnet: ensAddress,
+        query: {
+          addresses: {
+            testnet: ensAddress,
+          },
         },
       }),
     });