+++ packages/js/plugins/ethereum/src/__tests__/e2e.spec.ts
@@ -1,5 +1,5 @@
 import { ethereumPlugin } from "..";
-import * as Schema from "../w3";
+import * as Schema from "../query/w3-man";
 
 import { Web3ApiClient, defaultIpfsProviders } from "@web3api/client-js";
 import { ensPlugin } from "@web3api/ens-plugin-js";
@@ -69,8 +69,10 @@ describe("Ethereum Plugin", () => {
         {
           uri: "w3://ens/ens.web3api.eth",
           plugin: ensPlugin({
-            addresses: {
-              testnet: ensAddress
+            query: {
+              addresses: {
+                testnet: ensAddress
+              }
             }
           })
         }
@@ -162,7 +164,7 @@ describe("Ethereum Plugin", () => {
       expect(response.errors).toBeUndefined();
       expect(response.data?.callContractStatic).toBeDefined();
       expect(response.data?.callContractStatic.error).toBeTruthy();
-      expect(response.data?.callContractStatic.result).toBe("missing revert data in call exception");
+      expect(response.data?.callContractStatic.result).toContain("missing revert data in call exception");
     });
 
     it("getBalance", async () => {