+++ packages/js/plugins/filesystem/src/w3-man/index.ts
@@ -1,0 +1,3 @@
+export * from "./schema";
+export * from "./manifest";
+export * from "./plugin";