+++ packages/js/plugins/uts46/src/query/w3-man/types.ts
@@ -1,0 +1,21 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
+/* eslint-disable @typescript-eslint/naming-convention */
+
+export type UInt = number;
+export type UInt8 = number;
+export type UInt16 = number;
+export type UInt32 = number;
+export type Int = number;
+export type Int8 = number;
+export type Int16 = number;
+export type Int32 = number;
+export type Bytes = Uint8Array;
+export type BigInt = string;
+export type Json = string;
+export type String = string;
+export type Boolean = boolean;
+
+export interface ConvertResult {
+  IDN: string;
+  PC: string;
+}