+++ packages/test-cases/cases/bind/sanity/output/wasm-as/mutation/env.ts
@@ -3,3 +3,8 @@ import {
 } from "./MutationEnv";
 
 export let env: MutationEnv | null = null;
+
+export function requireEnv(): MutationEnv {
+  if (env == null) throw new Error("Undefined mutation env");
+  return env as MutationEnv;
+}