+++ packages/test-cases/cases/parse/sanity/output.ts
@@ -159,6 +159,16 @@ export const typeInfo: TypeInfo = {
           required: false,
         }),
         createScalarPropertyDefinition({
+          name: "bignumber",
+          type: "BigNumber",
+          required: true,
+        }),
+        createScalarPropertyDefinition({
+          name: "optBignumber",
+          type: "BigNumber",
+          required: false,
+        }),
+        createScalarPropertyDefinition({
           name: "json",
           type: "JSON",
           required: true,
@@ -459,20 +469,10 @@ export const typeInfo: TypeInfo = {
         {
           ...createMethodDefinition({
             type: "mutation",
-            name: "sanitizeMutationEnv",
-            return: createObjectPropertyDefinition({
-              name: "sanitizeMutationEnv",
-              type: "MutationEnv",
-              required: true,
-            }),
-            arguments: [
-              createObjectPropertyDefinition({
-                name: "env",
-                type: "MutationClientEnv",
-                required: true,
-              }),
-            ],
-          }),
+            name: "sanitizeEnv",
+            return: createObjectPropertyDefinition({ name: "sanitizeEnv", type: "MutationEnv", required: true }),
+            arguments: [createObjectPropertyDefinition({ name: "env", type: "MutationClientEnv", required: true })],
+          })
         },
       ],
     },
@@ -489,20 +489,10 @@ export const typeInfo: TypeInfo = {
         {
           ...createMethodDefinition({
             type: "query",
-            name: "sanitizeQueryEnv",
-            return: createObjectPropertyDefinition({
-              name: "sanitizeQueryEnv",
-              type: "QueryEnv",
-              required: true,
-            }),
-            arguments: [
-              createObjectPropertyDefinition({
-                name: "env",
-                type: "QueryClientEnv",
-                required: true,
-              }),
-            ],
-          }),
+            name: "sanitizeEnv",
+            return: createObjectPropertyDefinition({ name: "sanitizeEnv", type: "QueryEnv", required: true }),
+            arguments: [createObjectPropertyDefinition({ name: "env", type: "QueryClientEnv", required: true })],
+          })
         },
         {
           ...createMethodDefinition({