+++ packages/wasm/as/assembly/msgpack/Write.ts
@@ -1,5 +1,5 @@
 import { Nullable } from "./Nullable";
-import { BigInt } from "../math";
+import { BigInt, BigNumber } from "../math";
 import { Context } from "../debug";
 import { JSON } from "../json";
 
@@ -19,6 +19,7 @@ export abstract class Write {
   abstract writeBytesLength(length: u32): void;
   abstract writeBytes(value: ArrayBuffer): void;
   abstract writeBigInt(value: BigInt): void;
+  abstract writeBigNumber(value: BigNumber): void;
   abstract writeJSON(value: JSON.Value): void;
   abstract writeArrayLength(length: u32): void;
   abstract writeArray<T>(
@@ -49,6 +50,7 @@ export abstract class Write {
   abstract writeNullableString(value: string | null): void;
   abstract writeNullableBytes(value: ArrayBuffer | null): void;
   abstract writeNullableBigInt(value: BigInt | null): void;
+  abstract writeNullableBigNumber(value: BigNumber | null): void;
   abstract writeNullableJSON(value: JSON.Value | null): void;
   abstract writeNullableArray<T>(
     a: Array<T> | null,